requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
allure-pytest==2.15.0
allure-python-commons==2.15.0
aniso8601==10.0.1
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
black==25.9.0
blinker==1.9.0
boto3==1.38.16
botocore==1.38.16
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
ConfigArgParse==1.7.1
coverage==7.10.6
cryptography==44.0.2
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
execnet==2.1.1
executing==2.2.1
fastapi==0.115.12
firebase-admin==6.7.0
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
flask-restx==1.3.0
frozenlist==1.6.0
gevent==25.5.1
geventhttpclient==2.3.4
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.2.4
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.25
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
llama_cpp_python==0.3.9
locust==2.40.5
locust-cloud==1.27.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack==1.1.0
multidict==6.4.3
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.78.1
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.4
pytest-cov==7.0.0
pytest-html==3.2.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.15.1
pytest-rerunfailures==16.0.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-a2a==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9
s3transfer==0.12.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.1
structlog==25.4.0
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
zope.event==6.0
zope.interface==8.0
zstandard==0.23.0
